+++ src/components/Layout.jsx
@@ -1,4 +1,4 @@
-import Navbar from './Navbar';
+import Navbar from './navigation/Navbar';
 import { Outlet } from 'react-router-dom';
 
 function Layout() {
